Front-end and back-end parts of version saving were integrated

+++ Cortex/Cortex.Web/Models/Api/LayerModel.cs
@@ -26,9 +26,9 @@ namespace Cortex.Web.Models.Api
 
         public int NeuronsNumber { get; set; }
 
-        public double X { get; set; }
+        public int X { get; set; }
 
-        public double Y { get; set; }
+        public int Y { get; set; }
 
         public Layer ToDto()
         {
